+++ CHANGELOG.md
@@ -1,7 +1,11 @@
+# 1.0.x
+
+## Bug fixes
+- PIM-3444: Fix the enabled flag not published
+
 # 1.0.6 (2014-11-26)
 
 ## Bug fixes
-
 - PIM-3302: Published product page is not well displayed with project via the standard edition. To fix this bug, add `bundles/pimenterpriseui/css/pimee.less` to your stylesheets in `app/Resources/views/base.html.twig`
 
 # 1.0.5 (2014-11-13)